+++ CHANGELOG.md
@@ -12,3 +12,11 @@
 
 * Created backend storage construct
 
+## v0.4.0
+
+* Created networking construct
+
+## v0.5.0
+
+* Created certificate construct
+